pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-scraper"
version = "0.1.0"
description = "A Streamlit app for scraping Reddit posts based on search queries"
readme = "README.md"
authors = [{name = "Melissa Tan", email = "melissaxiaotan@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["reddit", "scraper", "streamlit", "selenium"]
dependencies = [
    "streamlit",
    "pandas",
    "praw",
    "selenium",
    "beautifulsoup4",
    "python-dotenv",
]
requires-python = ">=3.7"

[project.optional-dependencies]
dev = ["pytest", "black", "flake8"]

[project.urls]
Homepage = "https://github.com/my-moon-watcher/reddit-scraper"

[tool.setuptools]
packages = ["reddit_scraper"]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E266", "E501", "W503"]
max-complexity = 18
select = ["B","C","E","F","W","T4"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

